bluesky-bot/requirements.txt
atproto
psycopg2-binary
Pillow
fonttools
brotli
requests
rembg[cpu]
